requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
spotipy==2.19.0
urllib3==1.26.7
Werkzeug==2.0.2
